+++ engine/src/input/glfw/glfw_input.cpp
@@ -4,13 +4,13 @@
 #include "core/application.h"
 
 namespace Vulkyrie {
+
     [[nodiscard]] bool IsKeyPressed(const KeyCode key) {
         return glfwGetKey(static_cast<GLFWwindow *>(Application::GetSingleton().GetWindowHandle()), static_cast<u16>(key)) == GLFW_PRESS;
     }
 
     [[nodiscard]] bool IsMouseButtonPressed(const MouseButton button) {
-        return glfwGetMouseButton(static_cast<GLFWwindow *>(Application::GetSingleton().GetWindowHandle()), static_cast<u8>(button)) ==
-               GLFW_PRESS;
+        return glfwGetMouseButton(static_cast<GLFWwindow *>(Application::GetSingleton().GetWindowHandle()), static_cast<u8>(button)) == GLFW_PRESS;
     }
 
     [[nodiscard]] std::pair<f32, f32> GetMousePosition() {
@@ -18,7 +18,7 @@ namespace Vulkyrie {
 
         glfwGetCursorPos(static_cast<GLFWwindow *>(Application::GetSingleton().GetWindowHandle()), &xpos, &ypos);
 
-        return { (f32)xpos, (f32)ypos };
+        return { static_cast<f32>(xpos), static_cast<f32>(ypos) };
     }
 
     [[nodiscard]] f32 GetMouseX() {
@@ -26,7 +26,7 @@ namespace Vulkyrie {
 
         glfwGetCursorPos(static_cast<GLFWwindow *>(Application::GetSingleton().GetWindowHandle()), &xpos, nullptr);
 
-        return xpos;
+        return static_cast<f32>(xpos);
     }
 
     [[nodiscard]] f32 GetMouseY() {
@@ -34,6 +34,7 @@ namespace Vulkyrie {
 
         glfwGetCursorPos(static_cast<GLFWwindow *>(Application::GetSingleton().GetWindowHandle()), nullptr, &ypos);
 
-        return ypos;
+        return static_cast<f32>(ypos);
     }
+
 } // namespace Vulkyrie
